tidy_files/requirements.txt
# Aquaponics System v2.0 Dependencies
# 아쿠아포닉스 시스템 의존성 패키지

# Core Python packages
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
pyyaml>=6.0
pandas>=1.3.0
requests>=2.25.0

# Computer Vision & AI
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0

# Note: For Raspberry Pi or CPU-only systems, use:
# torch>=1.9.0+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
# torchvision>=0.10.0+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html

# Web Framework
flask>=2.0.0
flask-cors>=3.0.0

# MQTT Communication
paho-mqtt>=1.6.0

# Database
# sqlite3 is built-in to Python

# System monitoring
psutil>=5.8.0

# Data visualization (for dashboard)
plotly>=5.0.0
streamlit>=1.28.0

# Development & Debug
logging  # built-in
json     # built-in
datetime # built-in
uuid     # built-in
pathlib  # built-in
threading # built-in
queue    # built-in
time     # built-in

# Optional: Hardware-specific packages
# Uncomment based on your hardware:

# Raspberry Pi GPIO (uncomment for Raspberry Pi)
# RPi.GPIO>=0.7.0
# gpiozero>=1.6.0

# Camera support
# picamera>=1.13  # for Raspberry Pi camera
# v4l2py>=0.6.0   # for USB cameras

# Serial communication (for sensors)
# pyserial>=3.5

# Advanced analytics (optional)
# scikit-learn>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0

# Email notifications (optional)
# smtplib  # built-in
# email    # built-in

# Development tools (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=4.0.0
